backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
catboost==1.2.7
torch==2.4.0
pytorch-lightning==2.4.0
pytorch-forecasting==1.7.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
shap==0.46.0
scipy==1.13.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.8.0
redis==5.0.8
python-dotenv==1.0.0
umap-learn==0.5.6